+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- fix: 画像レイヤの const.sn.lay[lay].fore.width, height の値異常を修正
+
+
 ## [1.19.1](https://github.com/famibee/SKYNovel/compare/v1.19.0...v1.19.1) (2021-10-11)
 
 
@@ -24,7 +27,7 @@
 * タグ[page]追加。本やWeb漫画のように、既読ページを前後移動できる機能 ([ab6c328](https://github.com/famibee/SKYNovel/commit/ab6c3281f6a5ac1913a7dd5490d3d3740bf72963))
 
 - feat: タグ[page]追加。本やWeb漫画のように、既読ページを前後移動できる機能
-	- （レイヤの .page とページ移動の [page] の言葉混乱注意）
+	- （レイヤの fore/back ページ と、ページ移動の [page] の言葉で混乱しないよう注意）
 	- [p][s]など（[l][waitclick]は対象外）停止位置を記録し、前後できる
 	- ここでの【ページ】とは、【スクリプト先頭から一つめの[p]まで】、
 		【[p]の次トークン（文字やタグなど）から次の[p]まで】、を指す。